Modularize algolia autocomplete JS
