+++ algolia/autocomplete.php
@@ -90,131 +90,3 @@
 		<span class="empty-query">"{{ data.query }}"</span>
 	</div>
 </script>
-
-<script type="text/javascript">
-	jQuery(function () {
-		/* init Algolia client */
-		var client = algoliasearch(algolia.application_id, algolia.search_api_key);
-
-		/* setup default sources */
-		var sources = [];
-		jQuery.each(algolia.autocomplete.sources, function(i, config) {
-			sources.push({
-				source: algoliaAutocomplete.sources.hits(client.initIndex(config['index_name']), {
-					hitsPerPage: config['max_suggestions'],
-					attributesToSnippet: [
-						'content:10',
-						'title1:10',
-						'title2:10',
-						'title3:10',
-						'title4:10',
-						'title5:10',
-						'title6:10'
-					]
-				}),
-				templates: {
-					header: function(data, algoliaResponse) {
-						return wp.template('autocomplete-header')({
-							label: config['label'],
-              nbHits: algoliaResponse.nbHits,
-              resultsUrl: '#'
-						});
-					},
-          empty: wp.template('autocomplete-empty'),
-					suggestion: wp.template(config['tmpl_suggestion'])
-				}
-			});
-
-		});
-
-		/* Setup dropdown menus */
-		jQuery("input[name='s']:not('.no-autocomplete')").each(function(i) {
-			var $searchInput = jQuery(this);
-
-			var config = {
-				debug: algolia.debug,
-				hint: false,
-				openOnFocus: true,
-				templates: {
-          dropdownMenu: '#tmpl-dropdown-menu'
-        }
-			};
-
-			if(algolia.powered_by_enabled) {
-				config.templates.footer = wp.template('autocomplete-footer');
-			}
-
-			/* Instantiate autocomplete.js */
-			algoliaAutocomplete($searchInput[0], config, sources)
-			.on('autocomplete:selected', function(e, suggestion, datasetName) {
-				/* Redirect the user when we detect a suggestion selection. */
-				window.location.href = suggestion.permalink;
-			});
-
-			var $autocomplete = $searchInput.parent();
-
-			/* Remove autocomplete.js default inline input search styles. */
-			$autocomplete.removeAttr('style');
-
-			/* Configure tether */
-			var $menu = $autocomplete.find('.aa-dropdown-menu');
-			var config = {
-				element: $menu,
-				target: this,
-				attachment: 'top left',
-				targetAttachment: 'bottom left',
-				constraints: [
-					{
-						to: 'window',
-						attachment: 'none element'
-					}
-				]
-			};
-
-			/* This will make sure the dropdown is no longer part of the same container as */
-			/* the search input container. */
-			/* It ensures styles are not overridden and limits theme breaking. */
-			var tether = new Tether(config);
-			tether.on('update', function(item) {
-				/* todo: fix the inverse of this: https://github.com/HubSpot/tether/issues/182 */
-				if (item.attachment.left == 'right' && item.attachment.top == 'top' && item.targetAttachment.left == 'left' && item.targetAttachment.top == 'bottom') {
-					config.attachment = 'top right';
-					config.targetAttachment = 'bottom right';
-
-					tether.setOptions(config, false);
-				}
-			});
-			$searchInput.on('autocomplete:updated', function() {
-				tether.position();
-			});
-			$searchInput.on('autocomplete:opened', function() {
-				updateDropdownWidth();
-			});
-
-
-			/* Trick to ensure the autocomplete is always above all. */
-			$menu.css('z-index', '99999');
-
-			/* Makes dropdown match the input size. */
-			var dropdownMinWidth = jQuery('.layout-wrapper').first().outerWidth();
-			function updateDropdownWidth() {
-				var inputWidth = $searchInput.outerWidth();
-				if (inputWidth >= dropdownMinWidth) {
-					$menu.css('width', $searchInput.outerWidth());
-				} else {
-					$menu.css('width', dropdownMinWidth);
-				}
-				tether.position();
-			}
-			jQuery(window).on('resize', updateDropdownWidth);
-		});
-
-		/* This ensures that when the dropdown overflows the window, Thether can reposition it. */
-		jQuery('body').css('overflow-x', 'hidden');
-
-		jQuery(document).on("click", ".algolia-powered-by-link", function(e) {
-			e.preventDefault();
-			window.location = "https://www.algolia.com/?utm_source=WordPress&utm_medium=extension&utm_content=" + window.location.hostname + "&utm_campaign=poweredby";
-		});
-	});
-</script>
