+++ package.json
@@ -22,6 +22,7 @@
   "homepage": "https://github.com/telabotanica/wp-theme-telabotanica#readme",
   "private": true,
   "devDependencies": {
+    "algoliasearch": "^3.20.4",
     "autocomplete.js": "^0.24.2",
     "autoprefixer": "^6.7.0",
     "css-loader": "^0.25.0",
@@ -38,6 +39,7 @@
     "style-loader": "^0.13.1",
     "svg-sprite-loader": "0.0.31",
     "svg-url-loader": "^1.1.1",
+    "tether": "^1.4.0",
     "url-loader": "^0.5.7",
     "webpack": "^2.2.0"
   }
